Implementing the save and load functions
"Save board as" now allows you to save a board to the specified location
and "load board" allows you to load from the specified location.
Locations are specified via a JFileChooser.  Save board does not work
properly yet, as default location depends on where the program is being
run from.