Previous board no longer shows behind new board
When a new board is selected the previous one is no longer able to be
seen behind the new one.  A background coloured rectangle is drawn over
it before the new board is painted.  Also, size 14 boards have been
changed to size 13 as it is the proper size, 14 was a mistake on my
part.